patchbomb... again... added macalike menu plasmoid, changed tooltip look, move config to cmake for KDE support
git-svn-id: https://cloudcity.svn.sourceforge.net/svnroot/cloudcity@192 d23db356-9f35-0410-88be-f4158e15d4e5

+++ animator/hover.cpp
@@ -152,7 +152,7 @@ Hover::timerEvent(QTimerEvent * event)
       else { // fade out
          *step += HOVER_IN_STEP;
          widget->repaint();
-         if (*step > maxSteps-2) {
+         if ((uint)(*step) > maxSteps-2) {
 #if WOBBLE_HOVER
             if (widget->testAttribute(Qt::WA_UnderMouse))
                it.value().fadeIn = false;
